third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/reveal-execution-line-expected.txt

Tests that execution line is revealed and highlighted when debugger is paused.


Running: testRevealAndHighlightExecutionLine
Set timer for test function.
Execution line revealed and highlighted.
